@@ -1,57 +1,128 @@
-import { Button, Modal, ModalBody, ModalContent, ModalFooter, ModalHeader, Spacer, useDisclosure } from "@nextui-org/react";
-import { UserPlus } from "lucide-react";
+import { Button, Modal, ModalBody, ModalContent, ModalFooter, ModalHeader, Spacer } from "@nextui-org/react";
+import { LockKeyholeIcon } from "lucide-react";
 import Image from "next/image";
-import { useRouter } from "next/router";
-import { use, useEffect } from "react";
-import useAcceptInviteUtils from "@/hooks/useAcceptInviteUtils";
+import { useRouter } from "next/navigation";
+import useInviteUtils from "@/hooks/useInviteUtils";
+import AppInput from "../forms/AppInput";
+import { IInvite } from "@/types/Invite";
+import { FC, useState } from "react";
+import toast from "react-hot-toast";
+import { FormProvider, useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import z from "zod";
 
-const AcceptUserInviteModal = () => {
-	const { isOpen, onOpen, onOpenChange } = useDisclosure();
+const schema = z
+	.object({
+		password: z.string().min(8, {
+			message: "Password must be at least 8 characters long",
+		}),
+		confirmPassword: z.string().min(8, {
+			message: "Password must be at least 8 characters long",
+		}),
+	})
+	.refine((data) => data.password === data.confirmPassword, {
+		message: "Passwords do not match",
+		path: ["confirmPassword"],
+	});
+
+
+
+interface AcceptUserInviteModalProps {
+	isOpen: boolean;
+	setIsOpen: (value: boolean) => void;
+	onClose: () => void;
+	inviteInfo: IInvite;
+}
+
+const AcceptUserInviteModal: FC<AcceptUserInviteModalProps> = ({ isOpen, onClose, inviteInfo, setIsOpen }) => {
 	const router = useRouter();
-	const code = router.query.code;
-	const { acceptInvite, getInviteInfo } = useAcceptInviteUtils();
+	const { acceptInvite } = useInviteUtils();
+	const [loading, setLoading] = useState<boolean>(false);
+	const [error, setError] = useState<string | null>(null);
 
-	useEffect(() => {
-		const fetchInviteInfo = async () => {
-			if (code) {
-				const response = await getInviteInfo(code as string);
-				if (response.status === "success") {
-					console.log(response.data);
-				}
+	const formMethods = useForm<z.infer<typeof schema>>({
+		resolver: zodResolver(schema),
+		defaultValues: {
+			password: "",
+		},
+	});
+
+	const {
+		handleSubmit,
+		control,
+		reset,
+		formState: { errors },
+	} = formMethods;
+
+
+	const onSubmit = async (data: z.infer<typeof schema>) => {
+		setLoading(true);
+		const id = toast.loading("Accepting Invite ...");
+		try {
+			const response = await acceptInvite(inviteInfo.inviteCode, data.password);
+			if (response.status === "success") {
+				console.log(response.data);
+				toast.success("Invite Accepted Successfully", { id });
+				reset();
+				router.push("/auth/login");
+			} else {
+				toast.error(response.msg, { id });
+				setError(response.msg || "An error occurred. Please try again");
 			}
-		};
-		fetchInviteInfo();
-	}, [code]);
+		} catch (error) {
+			console.log(error?.response?.data?.msg);
+			toast.error(error?.response?.data?.msg || "An error occurred. Please try again", { id });
+			setError(error?.response?.data?.msg || "Please Contact your Admin");
+		} finally {
+			setLoading(false);
+			onClose();
+		}
+	};
 
 	return (
 		<>
-			<Button color="primary" startContent={<UserPlus className="w-4 h-4" />} 
-			  onPress={onOpen}>
-				AcceptInvite
-			</Button>
-
-			<Modal isOpen={isOpen} onOpenChange={onOpenChange} scrollBehavior="outside">
+			<Modal isOpen={isOpen} onOpenChange={setIsOpen} isDismissable={false}>
 				<ModalContent className="saastain" style={{ fontFamily: "Nunito" }}>
 					{(onClose) => (
-						<div className="p-6 bg-[#DEF7EC]">
-							<ModalHeader className="flex flex-col items-center">
-								<Image className="w-1/4 my-5" src={"/images/saastain_logo.svg"} alt="" width={100} height={100} />
-								<h2 className="text-lg font-semibold">
-									Join Company Name In Their <br /> Sustainability Journey
-								</h2>
-							</ModalHeader>
-							<ModalBody className="items-center">
-								<p className="mt-3">Saastain goes beyond conventional solutions, empowering you to not only navigate the complexities of emissions tracking but also to drive sustainable growth.</p>
-							</ModalBody>
-							<ModalFooter>
-								<Button color="danger" variant="bordered" onPress={onClose}>
-									Reject Invite
-								</Button>
-								<Button color="primary" onPress={onClose}>
-									Accept Invite
-								</Button>
-							</ModalFooter>
-						</div>
+						<FormProvider {...formMethods}>
+							<form onSubmit={handleSubmit(onSubmit)}>
+								<div className="p-6 bg-[#DEF7EC]">
+									<ModalHeader className="flex flex-col items-center">
+										<Image className="w-1/4 my-5" src={"/images/saastain_logo.svg"} alt="" width={100} height={100} />
+										<h2 className="text-sm">
+											You have been invited to join Company Name in their sustainability journey. <br />
+											Please set a password to join.
+										</h2>
+									</ModalHeader>
+									<ModalBody className="items-center">
+										<AppInput
+											label="Password"
+											type="password"
+											name="password"
+											control={control}
+											placeholder="Enter Password"
+											error={errors.password}
+											startContent={<LockKeyholeIcon className="text-sm text-default-400 pointer-events-none flex-shrink-0 mr-3" />}
+										/>
+										<Spacer y={4} />
+										<AppInput
+											name="confirmPassword"
+											placeholder="Confirm Password"
+											type="password"
+											control={control}
+											error={errors.confirmPassword}
+											startContent={<LockKeyholeIcon className="text-sm text-default-400 pointer-events-none flex-shrink-0 mr-3" />}
+										/>
+									</ModalBody>
+									<ModalFooter>
+										<Button variant="bordered" type="button">Cancel</Button>
+										<Button color="primary" type="submit" >
+											Submit
+										</Button>
+									</ModalFooter>
+								</div>
+							</form>
+						</FormProvider>
 					)}
 				</ModalContent>
 			</Modal>
