@@ -1,18 +1,63 @@
-import { useCallback } from "react";
-import { useApi } from "./useApi";
-import { IApiEndpoint, IApiResponse } from "@/types/Api";
-import { IInvite } from "@/types/Invite";
+import { IApiEndpoint, IApiResponse } from "@/types/Api"
+import { useApi } from "./useApi"
+import { useCallback } from "react"
+import { IInvite } from "@/types/Invite"
 
 const useInviteUtils = () => {
-	const { post } = useApi();
+    const {post, get} = useApi()
 
-	const inviteUserToCompany = useCallback(async (name: string, email: string, userRole: string) => {
+    const inviteUserToCompany = useCallback(async (name: string, email: string, userRole: string) => {
 		const resp = await post<IApiResponse<IInvite>>({ endpoint: IApiEndpoint.NEW_COMPANY_USER_INVITE, data: { name, email, userRole } });
 
 		return resp.data;
 	}, []);
 
-	return { inviteUserToCompany };
-};
+    const acceptInvite = useCallback (async (inviteCode: string, password:string) => {
+        const response = await post<IApiResponse>({
+            endpoint: IApiEndpoint.ACCEPT_INVITE,
+            data: {inviteCode, password},
+            checkAuth: false
+        })
+        return response.data
+    }, [])
+
+    const getInviteInfo = useCallback (async (queryVals: { code: string }) => {
+        const response = await get<IApiResponse<IInvite>>({
+            endpoint: IApiEndpoint.GET_INVITE_INFO,
+            queryParams: {code : queryVals.code},
+            checkAuth: false
+        })
+        return response.data
+    }, [])
+
+    const rejectInvite = useCallback (async (inviteCode: string) => {
+        const response = await post<IApiResponse>({
+            endpoint: IApiEndpoint.REJECT_INVITE,
+            data: {inviteCode},
+            checkAuth: false
+        })
+        return response.data
+    }, [])
+
+    const revokeInvite = useCallback (async (inviteCode: string) => {
+        const response = await post<IApiResponse>({
+            endpoint: IApiEndpoint.REVOKE_INVITE,
+            data: {inviteCode},
+            checkAuth: true
+        })
+        return response.data
+    }, [])
+
+
+    return{
+        acceptInvite,
+        getInviteInfo,
+        rejectInvite,
+        inviteUserToCompany,
+        revokeInvite
+    }
+
+
+}
 
 export default useInviteUtils;
