@@ -16,6 +16,7 @@ export type IApiResponse<T = any> = IApiSuccessResponse<T> | IApiErrorResponse;
 
 export const enum IApiEndpoint {
 	LOGIN = "auth/login",
+	VERIFY_EMAIL = "auth/verify-email",
 	REQUEST_PASSWORD_RESET = "auth/request-password-reset",
 	VERIFY_PASSWORD_RESET_TOKEN = "auth/verify-reset-password-token",
 	RESET_PASSWORD = "auth/reset-password",
@@ -23,9 +24,8 @@ export const enum IApiEndpoint {
 	GET_USERS = "users",
 	USER_PROFILE = "users/info",
 	CREATE_USER = "users",
+	CREATE_COMPANY_ADMIN = "users/company-admin",
 	CREATE_USER_BY_ADMIN = "users/admin/new-user",
-	ACCEPT_INVITE = "company/accept",
-	GET_INVITE_INFO = "info/code",
 	SUSPEND_ACCOUNT = "users/admin/suspend-user",
 	ACTIVATE_ACCOUNT = "users/admin/activate-user",
 	MARK_USER_ACCOUNT_AS_DELETED = "users/admin/mark-user-as-deleted",
@@ -35,10 +35,14 @@ export const enum IApiEndpoint {
 	UPDATE_USER_TO_SYSTEM_ADMIN = "users/admin/update-user-to-system-admin",
 	GET_COMPANY_USERS_PAGINATED = "users/company/fetch-with-filters",
 
+	CREATE_COMPANY = "company",
 	GET_COMPANIES = "company",
 	GET_COMPANY = "company/profile",
 	ADMIN_CREATE_COMPANY = "company/admin/create",
 	ADMIN_UPDATE_USER_WITH_COMPANY = "company/admin/add-user-to-company",
+	GET_COMPANY_BRANCHES = "company/branches/company",
+	CREATE_NEW_COMPANY_BRANCH = "company/branches/new",
+	DELETE_COMPANY_BRANCH = "company/branch/remove",
 
 	WAITLIST = "waitlist",
 
@@ -73,6 +77,29 @@ export const enum IApiEndpoint {
 	INVITES_COMPANY = "invites/company/all",
 	INVITES_COMPANY_PAGINATED = "invites/company/paginated",
 	NEW_COMPANY_USER_INVITE = "invites/company/new/user",
+	ACCEPT_INVITE = "invites/company/accept",
+	GET_INVITE_INFO = "invites/info/code",
+	REJECT_INVITE = "invites/company/reject",
+	REVOKE_INVITE = "invites/revoke",
+
+	// GET SCOPE TWO DATA ITEMS
+	GET_SCOPE_TWO_ELECTRICTY_DATA = "scope-two/get/electricity-data",
+
+	// GET SCOPE ONE DATA ITEMS
+	GET_SCOPE_ONE_VEHICLE_EMISSIONS_DATA = "scope-one/get/vehicles-emissions",
+	GET_SCOPE_ONE_PROCESS_EMISSIONS_DATA = "scope-one/get/process-emissions",
+	GET_SCOPE_ONE_FUGITIVE_EMISSIONS_DATA = "scope-one/get/fugitive-emissions",
+	GET_SCOPE_ONE_FUELS_EMISSIONS_DATA = "scope-one/get/fuel-emissions",
+
+	// UPDATE ACCOUNTING DATA
+	UPDATE_SCOPE_TWO_ELECTRICITY_DATA = "scope-two/update/electricity-data",
+	UPDATE_SCOPE_TWO_HEAT_AND_COOLING_DATA = "scope-two/update/heat-and-cooling-data",
+
+	// UPDATE SCOPE ONE ACCOUNTING DATA
+	UPDATE_SCOPE_ONE_VEHICLE_EMISSIONS_DATA = "scope-one/update/vehicles-emissions",
+	UPDATE_SCOPE_ONE_PROCESSING_EMISSIONS_DATA = "scope-one/update/process-emissions",
+	UPDATE_SCOPE_ONE_FUGITIVE_EMISSIONS_DATA = "scope-one/update/fugitive-emissions",
+	UPDATE_SCOPE_ONE_FUEL_EMISSIONS_DATA = "scope-one/update/fuel-emissions",
 }
 
 export interface IMethodParams {
